refactor: reorganize thumbnail constants and update song preview DTO to use specific visibility type
